create a specific Grist CSV Viewer section in README
Clean up language around the CSV Viewer a bit. It is still
inherently confusing, there are a lot of options and angles.

I changed CSV popups to default to single page view because
it feels like a better default for CSV, and it makes the
documentation simpler, one can reasonably leave most options
for a separate section without leaving the user with
something ugly.