update to near-latest grist-core
This updates to very nearly the latest version of Grist by adding
in little stubs to catch up with changes.

The very latest version of Grist started using `node:stream` imports
which confuses webpack prior to getting a chance to polyfill the
stream package - fixable I'm sure but I ran out of time.